Move all internal stuff to src/internal
