+++ packages/hardhat-chai-matchers/src/internal/properHex.ts
