+++ packages/hardhat-chai-matchers/src/panic.ts
@@ -1,0 +1,2 @@
+// re-export here so users can import from "@nomicfoundation/hardhat-chai-matchers/panic"
+export { PANIC_CODES } from "./internal/reverted/panic";
